zeaz/requirements.txt
fastapi==0.115.0
gunicorn==23.0.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
PyJWT==2.9.0
